fix: last commit before merge, move filters to profile.helpers.ts
