+++ src/app/(profile)/profile.helpers.ts
@@ -1,3 +1,4 @@
+import { type ProfileModel } from '@/app/(profile)/_models/profile.model'
 import {
   JobOfferFiltersEnum,
   type HourlyRateValue,
@@ -52,7 +53,6 @@ export const createFiltersObjFromSearchParams = (
   for (const [key, value] of searchParams.entries()) {
     filters[key as keyof SearchParamsFilters] = value.split(',')
   }
-  console.log(filters)
 
   return filters
 }
@@ -80,8 +80,8 @@ export const getHourlyRateDisplay = (
   currency: string,
   hourlyRateMax: number | null,
 ) => {
-  if (hourlyRateMin === null || hourlyRateMin < 10) {
-    return `Rate not specified`
+  if (hourlyRateMin === null || hourlyRateMin < 1) {
+    return null
   } else if (hourlyRateMin >= 350) {
     return `above ${hourlyRateMin} ${currency}/h`
   } else if (hourlyRateMin < 50) {
@@ -107,3 +107,84 @@ export const formatHourlyRateLabel = (
   if (max === null) return `> ${min} zł/h`
   return `${min} - ${max} zł/h`
 }
+
+export const filterByPosition =
+  (positionFilter: string[]) => (profile: ProfileModel) => {
+    if (!positionFilter.length) return true
+    return positionFilter.some(
+      (pos) => profile.position.toUpperCase() === pos.toUpperCase(),
+    )
+  }
+
+export const filterBySeniority =
+  (seniorityFilter: string[]) => (profile: ProfileModel) => {
+    if (!seniorityFilter.length) return true
+    return seniorityFilter.some(
+      (filter) => filter.toUpperCase() === profile.seniority.toUpperCase(),
+    )
+  }
+
+export const filterBySpecialization =
+  (specializations: string[]) => (profile: ProfileModel) => {
+    if (!specializations.length) return true
+    return specializations.some(
+      (filter) => filter.toUpperCase() === profile.position.toUpperCase(),
+    )
+  }
+
+// TODO: implement filter by location (cities and countreis should be implemented )
+export const filterByLocation =
+  (locationFilter: string[]) => (profile: ProfileModel) => {
+    if (!locationFilter.length) return true
+    return locationFilter.some((location) => profile.country === location)
+  }
+
+export const filterByTechnology =
+  (technologyFilter: string[]) => (profile: ProfileModel) => {
+    if (!technologyFilter.length) return true
+    return technologyFilter.some(
+      (techFilter) =>
+        !!profile.techStack.find(
+          (tech) => tech.name.toUpperCase() === techFilter.toUpperCase(),
+        ),
+    )
+  }
+
+export const filterByAvailability =
+  (availabilityFilter: string[]) => (profile: ProfileModel) => {
+    if (!availabilityFilter.length) return true
+    return availabilityFilter.some(
+      (availability) =>
+        !!profile.employmentTypes.find(
+          (employmentType) =>
+            employmentType.toUpperCase() === availability.toUpperCase(),
+        ),
+    )
+  }
+
+export const filterByFullName = (searchTermFilter: string) => {
+  return (profile: ProfileModel) => {
+    if (!searchTermFilter) return true
+    return profile.fullName
+      .toUpperCase()
+      .includes(searchTermFilter.toUpperCase())
+  }
+}
+
+export const filterBySalary =
+  (salaryFilter: string[]) => (profile: ProfileModel) => {
+    if (salaryFilter.length === 0) return true
+
+    return salaryFilter.some((salaryRange) => {
+      const [min, max] = salaryRange
+        .split('-')
+        .map((str) => (str === 'null' ? Infinity : Number(str)))
+      if (profile.hourlyRateMax === 0 && profile.hourlyRateMin === 350) {
+        return min >= 350
+      }
+      if (profile.hourlyRateMin === 0 && profile.hourlyRateMax === 0) {
+        return false
+      }
+      return profile.hourlyRateMin >= min && profile.hourlyRateMax <= max
+    })
+  }
